perf(onboarding): prerender the marketing homepage instead of reading the currency cookie on the server (#597)
